@@ -16,30 +16,12 @@ if TYPE_CHECKING:
     # pylint: disable=unused-import,ungrouped-imports
     from msrest import Deserializer, Serializer
 
-    from azure.core import PipelineClient
+    from azure.core import AsyncPipelineClient
 
-def _convert_request(request, files=None):
-    data = request.content if not files else None
-    request = HttpRequest(method=request.method, url=request.url, headers=request.headers, data=data)
-    if files:
-        request.set_formdata_body(files)
-    return request
-
-def _format_url_section(template, **kwargs):
-    components = template.split("/")
-    while components:
-        try:
-            return template.format(**kwargs)
-        except KeyError as key:
-            formatted_components = template.split("/")
-            components = [
-                c for c in formatted_components if "{}".format(key.args[0]) not in c
-            ]
-            template = "/".join(components)
 
 class MixinABC(ABC):
     """DO NOT use this class. It is for internal typing use only."""
-    _client: "PipelineClient"
+    _client: "AsyncPipelineClient"
     _config: DeviceUpdateConfiguration
     _serialize: "Serializer"
     _deserialize: "Deserializer"
