@@ -66,7 +66,9 @@ from ._device_update_enums import (
     Role,
     SKU,
 )
-
+from ._patch import __all__ as _patch_all
+from ._patch import *  # type: ignore # pylint: disable=unused-wildcard-import
+from ._patch import patch_sdk as _patch_sdk
 __all__ = [
     'Account',
     'AccountList',
@@ -125,3 +127,5 @@ __all__ = [
     'Role',
     'SKU',
 ]
+__all__.extend([p for p in _patch_all if p not in __all__])
+_patch_sdk()
